Release 0.0.3 Jest
Use RSpec style testing immediately in an Expo environment.  Snapshots give you a normalized and repeatable structural representation of the view hierarchy.  Mind blown !

Other than Webstorm, debugger configuration seems a bit of a mystery yet, not out of the box or I'm missing something thus far.  If I need a browser to drive tests that don't require it, then why not use an IDE ?  Sticking with Webstorm for now.